move toast-master to its own folder
